# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Artistic Hub API",
        "version": "0.0.1",
        "description": "Artistic Hub API: public catalog, customer accounts, cart, checkout and orders, and the admin API under /admin."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/customer/addresses/{address}/default": {
            "patch": {
                "operationId": "customer.v1.addresses.default",
                "description": "The previous default stops being the default. Only the customer's own addresses; any other id returns 404.",
                "summary": "Make an address the default",
                "tags": [
                    "AddressDefault"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`CustomerAddressResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerAddressResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "AddressLabel": {
                "type": "string",
                "description": "customer_addresses.label.\n",
                "enum": [
                    "home",
                    "work",
                    "other"
                ],
                "title": "AddressLabel"
            },
...
            "CustomerAddressResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Internal database id: shown for reference only; URLs and requests take public_id."
                    },
                    "public_id": {
                        "type": "string"
                    },
                    "label": {
                        "$ref": "#/components/schemas/AddressLabel"
                    },
                    "recipient_name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "address_line_1": {
                        "type": "string"
                    },
                    "address_line_2": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "postal_code": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "is_default": {
                        "type": "boolean",
                        "description": "Exactly one of the customer's addresses is the default."
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "public_id",
                    "label",
                    "recipient_name",
                    "phone",
                    "address_line_1",
                    "address_line_2",
                    "city",
                    "state",
                    "postal_code",
                    "country",
                    "is_default",
                    "created_at",
                    "updated_at"
                ],
                "title": "CustomerAddressResource"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}